==> analysis/04_Biodiversity.Rmd <==
---
title: "Biodiversity Analysis"
author: "Alicia McElwee"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: yes
    toc_float: yes
    collapsed: no
    smooth_scroll: yes
    toc_depth: 3
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      fig.path = "../figures/04_Biodiversity/") # Sends any figure output to this folder
```


# Setting the environment

-change knit directory to Project Directory

## Set seed
```{r set-seed}
set.seed(5318008)
```

## Load Libraries
```{r load-libraries}
pacman::p_load(devtools, phyloseq, tidyverse, dada2, iNEXT, patchwork, install = FALSE)

```

## Load in Data
```{r load-data}
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq


# Make a metadata dataframe
metadata_df <-
  raw_preprocessed_physeq %>%
  sample_data() %>%
  data.frame()

head(metadata_df) 

# Intuition Check
min(sample_sums(raw_preprocessed_physeq))

# Setting colors for the stations
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")



```

# Goals

1. Calculate the Hill Diversity of the samples.
2. Evaluate the rarefaction curves
3. Evaluate the Diversity values.
4. Make notes of specific samples and their seq depth. (flag samples with non-platuing rarefaction curves)

# Diversity Calculations with iNEXT

```{r calc-div}
# Prepare input data
iNEXT_input_df <-
  raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()

# Quick Check
dim(iNEXT_input_df) # First number is row which are ASVs, second is columns as number of samples

# Run iNEXT: calculate the Hill Numbers
# Note that: Species in ROW, Samples in COLUMNS
# Remember to set the seed!
#iNEXT_data <- iNEXT(iNEXT_input_df, 
#                    q = c(0,1,2), datatype = "abundance")

# Save the file
#save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")

```

# Evaluate the Diversity!
```{r load-div}
load("data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
typeof(iNEXT_data) # A list type that is less stringent than phyloseq object, this is S3 object while phyloseq is S4, S3 doesn't require all data to be anchored to eachother by common tag
```


# Plot Diversity
```{r plot-rarefaction}
# Prepare colors
color_df <- 
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()

# Check
head(color_df)

# Rename column
colnames(color_df)[1] <- "names"
# Check
head(color_df)

# Make a helper dataframe for plotting with colors
iNEXT_color_df <-
  color_df %>%
  # Fix names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-")) %>%
  mutate(names = gsub(names, pattern = "X", replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station")
  

```


# Plot Rarefaction with `ggiNEXT`

```{r ggiNEXT, fig.width=8, fig.height=3.5}
# Plot rarefaction
# rarefaction/extrapolation curve, type = 1

# Order q:
  # 0 = richnes/ Total number taxa
  # 1 = Exponential Shannon/ number of "common" taxa
  # 2 = inverse Simpson/ Number of "dominant" taxa

ggiNEXT(iNEXT_data, type = 1, facet.var= "Order.q") +
  facet_wrap(~Order.q, scales = "fixed") +
  scale_color_manual(values = iNEXT_color_df$station_colors, guide = FALSE) +
  scale_fill_manual(values = iNEXT_color_df$station_colors, guide = FALSE) +
  scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)), guide = FALSE) +
  theme(legend.position = "none")

```

# Manually plot diversity


## Rarefaction curve
```{r iNEXT-manual}

iNEXT_manual_df <-
  iNEXT_data$iNextEst$size_based %>% # $ in S3 object subsets different lists
  # Change assemblage to names
  dplyr::rename(names = Assemblage) %>% 
  # Changing sample name format to correct names format
  mutate(names = gsub(names, pattern = "[.]", replace = "-"), 
         names = gsub(names, pattern = "X", replace = "")) %>% 
  # join with metadata
  left_join(., metadata_df, by = "names") %>% 
  # Add colors to data frane
  left_join(., data.frame(station_colors = station_colors,
                          station = names(station_colors)), by = "station") 
  
# Plot it - Rarefaction Curve
 iNEXT_manual_df %>%
   # Filter out rows that are calculated by rarefaction from iNEXT
   dplyr::filter(Method == "Observed") %>%
   # Make the actual plot 
   ggplot(aes(x = m, y=qD, group_by = names, color = station)) +
   geom_line() +
   facet_grid(Order.q ~ station, scales = "fixed") +
   xlab("Number of Sequences") +
   ylab("Species Diversity") +
   scale_color_manual(values = station_colors) +
   theme(legend.position = "bottom")



```


# Environmental Variable Check

```{r environmental-pairs, fig.width=8, fig.height=8}
head(metadata_df)

# Pull out environmental variables
env_df <-
  metadata_df %>%
  dplyr::select(names, water_tempC:DO_mgL)

head(env_df)

# plot the correlations between environmental variables
pairs(dplyr::select(env_df, -names), upper.panel = NULL)

```

# Diversity vs Salinity

```{r div-vs-sal, fig.height=3.5, fig.width = 6}
# PSU = practical salinity unit
# 1 PSU = 1 gram of salt per 1000 grams of water
# 1 PSU = 1 ppt (part per thousand)
# Freshwater is typically between 0 and 0.5
# Brackish water is between 0.5 and 30
# Marine is anything greater than 30

iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x = salinity_psu, y = qD)) +
  facet_wrap(.~Order.q, scales = "free") +
  geom_point(aes(color = station)) +
  stat_smooth(method = "lm", formula = y ~poly(x,2)) +
  theme(legend.position = "bottom") +
  scale_color_manual(values = station_colors) +
  labs(x = "Salinity (PSU)", y = "Number of ASVs")
  
```

# Session Information

```{r session-info}
# Ensure reproducibility by telling software versions
devtools::session_info()
```
